+++ core/init.php
@@ -8,26 +8,32 @@
      *
     **/
 
-	define('WOK_MAJOR_VERSION', 0); // Major version
-	define('WOK_MINOR_VERSION', 3); // Minor version
-	define('WOK_RELEASE_VERSION', 1); // Release version
-	define('WOK_EXTRA_VERSION', 'building'); // Extra version
-	define('WOK_VERSION', WOK_MAJOR_VERSION.'.'.WOK_MINOR_VERSION.':'.WOK_RELEASE_VERSION); // Full version (without extra)
+	const WOK_MAJOR_VERSION        = 0; // Major version
+	const WOK_MINOR_VERSION        = 4; // Minor version
+	const WOK_RELEASE_VERSION      = 1; // Release version
+	const WOK_EXTRA_RELEASE        = 'building'; // Extra version
+    
+    // Define full WOK version (without extra release)
+	define('WOK_VERSION', WOK_MAJOR_VERSION.'.'.WOK_MINOR_VERSION.':'.WOK_RELEASE_VERSION);
 	
-
+    
     /*
      * The following lines will define default path of essential tools.
      * We suggest you to let them as they are for a better compatibility.
     **/
-	define('SYSTEM_ROOT', dirname(dirname(__FILE__))); // Define absolute root path
-    define('PATH_CORE', '/core'); // Core path
-	define('PATH_LIBS', '/libraries'); // Libraries path
-    define('PATH_DATA', '/data'); // Data's directory path
-    define('PATH_LOCALES', '/languages'); // Languages' files directory
-	define('PATH_TEMPLATES', '/templates'); // Template's directory path
-    define('PATH_FILES', '/files'); // Files' directory
-    define('PATH_TMP_FILES', PATH_FILES.'/tmp'); // Temporary files' directory
 
+    // Define absolute root path ( PHP 5.3+ use dirname(__DIR__) )
+	define('SYSTEM_ROOT', dirname(dirname(__FILE__))); 
+
+    const PATH_CORE             = '/core'; // Core path
+	const PATH_LIBS             = '/libraries'; // Libraries path
+    const PATH_DATA             = '/data'; // Data's directory path
+    const PATH_LOCALES          = '/languages'; // Languages' files directory
+    const PATH_APIS             = '/api'; // API's path
+	const PATH_TEMPLATES        = '/templates'; // Template's directory path
+    const PATH_FILES            = '/files'; // Files' directory
+    const PATH_TMP_FILES        = '/files/tmp'; // Temporary files' directory
+    const PATH_LOGS             = '/logs'; // PHP logs directory
 
     /*
      * Once we have pathes, we can call essential libraries.
@@ -41,42 +47,38 @@
          * All right ! Settings file exists.
          * We can load settings and required libraries
         **/
-        require_once SYSTEM_ROOT.PATH_CORE."/settings.php"; // Framework settings
-        require_once SYSTEM_ROOT.PATH_CORE."/utilities.php"; // Framework fonctions
-        require_once SYSTEM_ROOT.PATH_CORE."/compatibility.php"; // PHP compatibility functions
-        require_once SYSTEM_ROOT.PATH_CORE."/treatments.php"; // Treatments functions
-        require_once SYSTEM_ROOT.PATH_CORE."/timezones.php"; // Timezones code / name
-        require_once SYSTEM_ROOT.PATH_CORE."/session.php"; // Session class
-        require_once SYSTEM_ROOT.PATH_CORE."/locales.php"; // Locales class
-        require_once SYSTEM_ROOT.PATH_CORE."/file.php"; // File class
-        require_once SYSTEM_ROOT.PATH_CORE."/mail.php"; // Mail class
-
+        require_once SYSTEM_ROOT.PATH_CORE . '/settings.php'; // Framework settings
+        
+        require_once SYSTEM_ROOT.PATH_CORE . '/console.php'; // Framework logger
+        require_once SYSTEM_ROOT.PATH_CORE . '/utilities.php'; // Framework fonctions
+        require_once SYSTEM_ROOT.PATH_CORE . '/compatibility.php'; // PHP compatibility functions
+        require_once SYSTEM_ROOT.PATH_CORE . '/treatments.php'; // Treatments functions
+        require_once SYSTEM_ROOT.PATH_CORE . '/timezones.php'; // Timezones code / name
+        require_once SYSTEM_ROOT.PATH_CORE . '/session.php'; // Session class
+        require_once SYSTEM_ROOT.PATH_CORE . '/locales.php'; // Locales class
+        require_once SYSTEM_ROOT.PATH_CORE . '/file.php'; // File class
+        require_once SYSTEM_ROOT.PATH_CORE . '/mail.php'; // Mail class
 
         /**
          * Particular cases which require some adjustements or conditions
         **/
-        if(!function_exists('json_decode') && !function_exists('json_encode')):
-			require_once SYSTEM_ROOT.PATH_CORE."/json.php"; // JSON functions
-		endif;
-      
+        if(!function_exists('json_decode') && !function_exists('json_encode'))
+			require_once SYSTEM_ROOT.PATH_CORE . "/json.php"; // JSON functions
 
         /**
          * Start and send every required headers.
         **/
         if(!headers_sent()):
             session_start(); // Start sessions
-            header("Content-Type: text/html; charset=utf-8"); // Send charset
             @date_default_timezone_set(SYSTEM_TIMEZONE); // Define date timezone
         endif;
-        
-
+    
         /**
          * Once everything is fine loaded, we call the options file.
          * This one will be used to add your own stuffs.
         **/
-        if(file_exists(SYSTEM_ROOT.'/core/options.php')):
-			require_once(SYSTEM_ROOT.'/core/options.php');
-		endif;
+        if(file_exists(SYSTEM_ROOT.PATH_CORE.'/options.php'))
+			require_once(SYSTEM_ROOT.PATH_CORE.'/options.php');
         
     endif;