+++ core/settings.php
@@ -21,13 +21,19 @@
 	define('SYSTEM_TIMEZONE', 'Europe/Paris'); // System timezone
     define('SYSTEM_ACCEPT_LANGUAGES', 'en_EN,fr_FR'); // System accepted languages
 
+    
+    /**
+     * System logs
+    **/
+    define('CONSOLE_LOGS_RECEIVER', 'sebasalex@free.fr');
+    
 
     /**
      *	Security constants
      * You can use it to add salt in sessions, tokens or cookies' name/value
     **/
-	define('SESSION_CRYPT', 'e0266047bd12e95e0b410581a9a8a748eb8952e5'); // sha1(uniqid('sess_', true));
-	define('TOKEN_SALT', 'f27e108aa1b1a92e8b7c2ac6437aa3a1ec4097ee'); // sha1(uniqid('tok_', true));
-	define('COOKIE_CRYPT', 'b873ae4dcfb7e109b949d39c68bec689eab9572f');  // sha1(uniqid('cook_', true));
+	define('SESSION_CRYPT', '1eca304156306396a85875ba9b96124335a5614a'); // sha1(uniqid('sess_', true));
+	define('TOKEN_SALT', '1a61969b002ab8a3ee2a050ebd8d28c5a043a17b'); // sha1(uniqid('tok_', true));
+	define('COOKIE_CRYPT', '3e400e8c954584628a156305cf79b9886a0fcafb');  // sha1(uniqid('cook_', true));
 	
 ?>