+++ languages/fr_FR/sources/default.php
@@ -1,0 +1,14 @@
+<?php
+
+    $default = array(
+        'buttons' => array(
+            'next' => "Suivant",
+            'previous' => "Précédent",
+            'more' => "Dites m'en plus",
+            'donotcare' => 'Terminer la visite',
+            'finish' => "Terminer",
+        ),
+        
+    );
+
+?>