+++ templates/inc/headers.php
@@ -1,5 +1,9 @@
 <?php get_library('bootstrap'); ?>
 <?php get_library('jquery'); ?>
-<?php get_library('prism'); ?>
 
-<link href="<?php echo path(PATH_TEMPLATES."/style/default.css"); ?>" rel="stylesheet" type="text/css">
+<link href="<?php echo path(PATH_TEMPLATES."/css/default.css"); ?>" rel="stylesheet" type="text/css">
+<link href="<?php echo path(PATH_TEMPLATES."/css/bootstrap.min.css"); ?>" rel="stylesheet" type="text/css">
+<link href="<?php echo path(PATH_TEMPLATES."/css/bootstrap-responsive.min.css"); ?>" rel="stylesheet" type="text/css">
+
+<script src="<?php echo path(PATH_TEMPLATES.'/js/jquery-1.10.2.js'); ?>"></script>
+<script src="<?php echo path(PATH_TEMPLATES.'/js/bootstrap.min.js'); ?>"></script>